modify: rename NavbarLanding and add footer
