+++ frontend/app/providers/QueryProvider.js
@@ -28,7 +28,7 @@ export default function QueryProvider({ children }) {
   return (
     <QueryClientProvider client={queryClient}>
       {children}
-      <ReactQueryDevtools initialIsOpen={false} /> *
+      <ReactQueryDevtools initialIsOpen={false} />
     </QueryClientProvider>
   );
 }
