+++ frontend/components/LandingNavbar.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const LandingNavbar = () => {
+  return (
+    <div>
+      LandingNavbar
+    </div>
+  )
+}
+
+export default LandingNavbar
